[Feat] #33 - 김가현 코드리뷰 반영
